refactor: move common features in ConfluxTestFrameworkForContract to ConfluxTestFramework
