+++ tests/test_framework/test_framework.py
@@ -3,7 +3,8 @@
 # Distributed under the MIT software license, see the accompanying
 # file COPYING or http://www.opensource.org/licenses/mit-license.php.
 """Base class for RPC testing."""
-from typing import List, Literal, Union
+from typing import List, Literal, Union, Any, Type, cast
+from functools import cached_property
 from conflux.config import DEFAULT_PY_TEST_CHAIN_ID
 from conflux.messages import Transactions
 from conflux.rpc import RpcClient, default_config
@@ -20,6 +21,15 @@ import time
 from typing import Union
 import random
 
+from conflux_web3 import Web3 as CWeb3
+from conflux_web3.middleware.base import ConfluxWeb3Middleware
+from conflux_web3._utils.rpc_abi import (
+    RPC
+)
+from conflux_web3.contract import ConfluxContract
+from web3 import Web3
+from web3.middleware.signing import SignAndSendRawMiddlewareBuilder
+from web3.types import RPCEndpoint
 from cfx_account import Account as CoreAccount
 from eth_account import Account
 
@@ -45,7 +55,10 @@ from .util import (
     sync_blocks,
     sync_mempools,
     wait_until,
+    load_contract_metadata,
+    InternalContractName,
 )
+from .block_gen_thread import BlockGenThread
 
 class TestStatus(Enum):
     PASSED = 1
@@ -57,6 +70,7 @@ TEST_EXIT_PASSED = 0
 TEST_EXIT_FAILED = 1
 TEST_EXIT_SKIPPED = 77
 
+Web3NotSetupError = ValueError("w3 is not initialized, please call self.setup_w3() first")
 
 class ConfluxTestFramework:
     """Base class for a bitcoin test script.
@@ -73,6 +87,10 @@ class ConfluxTestFramework:
     The __init__() and main() methods should not be overridden.
 
     This class also contains various public and private helper methods."""
+
+    _cw3: Union[CWeb3, None] = None
+    _ew3: Union[Web3, None] = None
+    num_nodes: int = 0
 
     def __init__(self):
         """Sets test framework defaults. Do not override this method. Instead, override the set_test_params() method"""
@@ -111,7 +129,61 @@ class ConfluxTestFramework:
                 self.evm_secrets.append(Account.create().key.hex())
             if "core" in space or "core" == space:
                 self.core_secrets.append(Account.create().key.hex())
-            
+
+    @cached_property
+    def client(self) -> RpcClient:
+        """Get the RPC client, using the first node.
+        The RPC client is 
+
+        Returns:
+            RpcClient: used to send RPC requests to the node.
+                For example, self.client.cfx_getBalance(...) or self.client.eth_getBalance(...)
+                It should be noticed that the parameters are usually not formatted. 
+                Certain methods also provide formatted parameters, for example, self.client.epoch_number().
+                Please check the source code for more details.
+        """
+        return RpcClient(self.nodes[0])
+
+    @property
+    def cw3(self) -> CWeb3:
+        """Get the Conflux Web3 instance, initialized by self.setup_w3().
+
+        Raises:
+            Web3NotSetupError: If the Web3 instance is not initialized.
+
+        Returns:
+            CWeb3: The Conflux Web3 instance.
+        """
+        if self._cw3 is None:
+            raise Web3NotSetupError
+        return self._cw3
+    
+    @property
+    def ew3(self) -> Web3:
+        """Get the EVM Web3 instance, initialized by self.setup_w3().
+
+        Raises:
+            Web3NotSetupError: If the Web3 instance is not initialized.
+
+        Returns:
+            Web3: The EVM Web3 instance.
+        """
+        if self._ew3 is None:
+            raise Web3NotSetupError
+        return self._ew3
+    
+    @property
+    def cfx(self):
+        if self._cw3 is None:
+            raise Web3NotSetupError
+        return self._cw3.cfx
+    
+    @property
+    def eth(self):
+        if self._ew3 is None:
+            raise Web3NotSetupError
+        return self._ew3.eth
+   
     @property
     def core_accounts(self):
         return [CoreAccount.from_key(key, network_id=DEFAULT_PY_TEST_CHAIN_ID) for key in self.core_secrets]
@@ -245,7 +317,7 @@ class ConfluxTestFramework:
             self.setup_chain()
             self.setup_network()
             self.before_test()
-            self.run_test()
+            self.run_test()  # type: ignore
             success = TestStatus.PASSED
         except JSONRPCException as e:
             self.log.exception("JSONRPC error")
@@ -345,6 +417,36 @@ class ConfluxTestFramework:
         """Override this method to customize test node setup"""
         self.add_nodes(self.num_nodes, genesis_nodes=genesis_nodes, binary=binary, is_consortium=is_consortium)
         self.start_nodes()
+
+    def setup_w3(self):
+        """Setup w3 and ew3 for EVM and Conflux.
+        This method should be called before any test.
+        Use self.w3 and self.ew3 to access the web3 instances.
+        Use self.cfx and self.eth to access the Conflux and EVM RPC clients.
+        """
+        client = RpcClient(self.nodes[0])
+        log = self.log
+        self._cw3 = CWeb3(CWeb3.HTTPProvider(f'http://{self.nodes[0].ip}:{self.nodes[0].rpcport}/'))
+        self._ew3 = Web3(Web3.HTTPProvider(f'http://{self.nodes[0].ip}:{self.nodes[0].ethrpcport}/'))
+        self.cw3.wallet.add_accounts(self.core_accounts)
+        self.cw3.cfx.default_account = self.core_accounts[0].address
+        
+        self.ew3.middleware_onion.add(SignAndSendRawMiddlewareBuilder.build(self.evm_secrets)) # type: ignore
+        self.eth.default_account = self.evm_accounts[0].address
+        
+        class TestNodeMiddleware(ConfluxWeb3Middleware):
+            def request_processor(self, method: RPCEndpoint, params: Any) -> Any:
+                if method == RPC.cfx_sendRawTransaction or method == RPC.cfx_sendTransaction:
+                    client.node.wait_for_phase(["NormalSyncPhase"])
+                return super().request_processor(method, params)
+
+            def response_processor(self, method: RPCEndpoint, response: Any):
+                if method == RPC.cfx_getTransactionReceipt:
+                    if "result" in response and response["result"] is None:
+                        log.debug("Auto generate 5 blocks because did not get tx receipt")
+                        client.generate_blocks_to_state(num_txs=1)  # why num_txs=1?
+                return response
+        self.cw3.middleware_onion.add(TestNodeMiddleware)
 
     def add_nodes(self, num_nodes, genesis_nodes=None, rpchost=None, binary=None, auto_recovery=False,
                   recovery_timeout=30, is_consortium=True):
@@ -527,6 +629,56 @@ class ConfluxTestFramework:
                 if int(i["outcomeStatus"], 0) != 0:
                     raise AssertionError("Receipt states the execution failes: {}".format(i))
         return receipts    
+
+    def start_block_gen(self):
+        BlockGenThread(self.nodes, self.log).start()
+        
+    def enable_max_priority_fee_per_gas(self):
+        # enable cfx_maxPriorityFeePerGas
+        # or Error(Epoch number larger than the current pivot chain tip) would be raised
+        self.client.generate_blocks_to_state(num_txs=1)
+
+    def cfx_contract(self, name) -> Type[ConfluxContract]:
+        metadata = load_contract_metadata(name)
+        return self.cfx.contract(
+            abi=metadata["abi"], bytecode=metadata["bytecode"])
+
+    def internal_contract(self, name: InternalContractName):
+        return self.cfx.contract(name=name, with_deployment_info=True)
+
+    def deploy_contract(self, name, transact_args = {}) -> ConfluxContract:
+        tx_hash = self.cfx_contract(name).constructor().transact(transact_args)
+        receipt = tx_hash.executed(timeout=30)
+        return self.cfx_contract(name)(cast(str, receipt["contractCreated"]))
+
+    def deploy_create2(self):
+        self.create2factory = self.deploy_contract("Create2Factory")
+        self.client.generate_blocks(5)
+    
+    def deploy_contract_2(self, name, seed, *args, **kwargs) -> ConfluxContract:
+        if self.create2factory is None:
+            raise Exception("Create2Factory is not deployed")
+        contract_factory = self.cfx_contract(name)
+        deploy_code = contract_factory.constructor(*args, **kwargs)._build_transaction()["data"]
+        dest_address = self.create2factory.functions.callCreate2(seed, deploy_code).call()
+        self.create2factory.functions.callCreate2(seed, deploy_code).transact().executed()
+        return contract_factory(dest_address)
+
+    def cfx_transfer(self, receiver, value=None, gas_price=1, priv_key=None, decimals: int = 18, nonce = None, execute: bool = True):
+        if value is not None:
+            value = int(value * (10**decimals))
+        else:
+            value = 0
+
+        tx = self.client.new_tx(
+            receiver=receiver, gas_price=gas_price, priv_key=priv_key, value=value, nonce=nonce)
+        self.client.send_tx(tx, execute)
+        if execute:
+            self.wait_for_tx([tx], True)
+            receipt = self.client.get_transaction_receipt(tx.hash_hex())
+            return receipt
+        else:
+            return tx.hash_hex()
 
 class SkipTest(Exception):
     """This exception is raised to skip a test"""
